@@ -15,7 +15,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 0.0.9+1
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
@@ -30,8 +30,13 @@ dependencies:
   cupertino_icons: ^0.1.3
   http: ^0.12.1
   shared_preferences: ^0.5.7+1
-  google_maps_flutter: ^0.4.0
+  google_maps_flutter: ^0.5.27+2
   fancy_bottom_navigation: ^0.3.2
+  rounded_floating_app_bar: ^0.1.0
+  flutter_bloc: ^4.0.0
+  equatable: ^1.1.1
+  rxdart: ^0.24.1
+  google_fonts: ^0.3.9
 
 dev_dependencies:
   flutter_test:
