@@ -6,8 +6,9 @@ class Token implements BaseModel {
   final String tokenType;
   int expiresIn;
   final String refreshToken;
+  int createdTimestamp;
 
-  Token({ @required this.accessToken, @required this.tokenType, @required this.expiresIn, @required this.refreshToken });
+  Token({ @required this.accessToken, @required this.tokenType, @required this.expiresIn, @required this.refreshToken, @required this.createdTimestamp });
 
   factory Token.fromJson(Map<String, dynamic> json) {
     if (null == json["access_token"]) {
@@ -17,7 +18,8 @@ class Token implements BaseModel {
       accessToken: json["access_token"] as String, 
       tokenType: json["token_type"] as String,
       expiresIn: json["expires_in"] as int,
-      refreshToken: json["refresh_token"] as String
+      refreshToken: json["refresh_token"] as String,
+      createdTimestamp: (json["createdTimestamp"] != null) ? json["createdTimestamp"] : DateTime.now().millisecondsSinceEpoch
     );
   }
 
@@ -26,7 +28,8 @@ class Token implements BaseModel {
       "access_token": this.accessToken,
       "token_type": this.tokenType,
       "expires_in": this.expiresIn,
-      "refresh_token": this.refreshToken
+      "refresh_token": this.refreshToken,
+      "createdTimestamp": this.createdTimestamp
     };
   }
 
