@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:right_spot/view/home/home_view.dart';
-import 'package:right_spot/view/login_view.dart';
 
 void main() {
   runApp(MyApp());
@@ -10,11 +10,15 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
     return MaterialApp(
       title: 'Right Spot',
       theme: ThemeData(
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
+        textTheme:GoogleFonts.latoTextTheme(textTheme).copyWith(
+          bodyText2: GoogleFonts.montserrat(textStyle: textTheme.bodyText2),
+        ),
       ),
       home: HomeView()
     );
